+++ packages/plugin-interface/src/PluginDataFixtureInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Shopsys\Plugin;
+
+interface PluginDataFixtureInterface
+{
+    /**
+     * Loads plugin demo data
+     */
+    public function load();
+}