+++ packages/plugin-interface/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - This Changelog (@vitek-rostislav)
+- interface for loading plugin's demo data (@MattCzerner)
+    - `PluginDataFixtureInterface`
 
 ## 0.1.0 - 2017-08-04
 ### Added